Refactor, seperate out test_views.
